refactor modellist
