+++ CommonUI/src/Components/ModelList/ModelList.tsx
@@ -1,91 +1,190 @@
-
+import React, { ReactElement, useEffect, useState } from 'react';
+import Query from '../../Utils/ModelAPI/Query';
 import BaseModel from 'Common/Models/BaseModel';
-import Typeof from 'Common/Types/Typeof';
-import React, { ReactElement } from 'react';
+import ErrorMessage from '../ErrorMessage/ErrorMessage';
+import ComponentLoader from '../ComponentLoader/ComponentLoader';
+import ModelAPI, { ListResult } from '../../Utils/ModelAPI/ModelAPI';
+import { LIMIT_PER_PROJECT } from 'Common/Types/Database/LimitMax';
+import Select from '../../Utils/ModelAPI/Select';
+import HTTPErrorResponse from 'Common/Types/API/HTTPErrorResponse';
+import Input from '../Input/Input';
+import StaicModelList from '../ModelList/StaticModelList';
+import WorkflowVariable from 'Model/Models/WorkflowVariable';
 
 export interface ComponentProps<TBaseModel extends BaseModel> {
-    list: Array<TBaseModel>;
-    headerField?: string | ((item: TBaseModel) => ReactElement) | undefined;
-    descriptionField: string;
-    selectedItems: Array<TBaseModel>;
-    onClick: (item: TBaseModel) => void;
+    query?: Query<TBaseModel>;
+    modelType: { new(): TBaseModel };
     titleField: string;
+    isSearchEnabled?: boolean | undefined;
+    descriptionField?: string | undefined;
+    selectMultiple?: boolean | undefined;
+    select: Select<TBaseModel>;
+    noItemsMessage: string;
+    headerField?: string | ((item: TBaseModel) => ReactElement) | undefined;
+    onSelectChange: (list: Array<TBaseModel>)=> void;
 }
 
 const ModelList: Function = <TBaseModel extends BaseModel>(
     props: ComponentProps<TBaseModel>
 ): ReactElement => {
+    const [selectedList, setSelectedList] = useState<Array<TBaseModel>>([]);
+    const [modelList, setModalList] = useState<Array<TBaseModel>>([]);
+    const [error, setError] = useState<string>('');
+    const [isLoading, setIsLoading] = useState<boolean>(false);
+    const [searchedList, setSearchedList] = useState<Array<TBaseModel>>([]);
+    const [searchText, setSearchText] = useState<string>('');
 
+
+    useEffect(()=> {
+        props.onSelectChange(selectedList);
+    }, [selectedList])
+
+
+    useEffect(() => {
+        fetchItems().catch();
+    }, []);
+
+    useEffect(() => {
+        if (!props.isSearchEnabled) {
+            setSearchedList([...modelList]);
+        }
+    }, [props.isSearchEnabled, modelList]);
+
+    const fetchItems: Function = async () => {
+        setError('');
+        setIsLoading(true);
+
+        try {
+            const listResult: ListResult<TBaseModel> =
+                await ModelAPI.getList<TBaseModel>(
+                    props.modelType,
+                    {
+                        ...props.query,
+                    },
+                    LIMIT_PER_PROJECT,
+                    0,
+                    props.select,
+                    {},
+                    {},
+                    {}
+                );
+
+            setModalList(listResult.data);
+        } catch (err) {
+            try {
+                setError(
+                    (err as HTTPErrorResponse).message ||
+                    'Server Error. Please try again'
+                );
+            } catch (e) {
+                setError('Server Error. Please try again');
+            }
+        }
+
+        setIsLoading(false);
+    };
+
+    useEffect(() => {
+        if (!searchText) {
+            setSearchedList([...modelList]);
+        } else {
+            // search
+
+            setSearchedList(
+                [...modelList].filter((model: TBaseModel): boolean => {
+                    const includedInSearch: boolean = (
+                        model.getValue(props.titleField) as string
+                    )
+                        .toLowerCase()
+                        .includes(searchText);
+
+                    if (!includedInSearch && props.descriptionField) {
+                        return (
+                            model.getValue(props.descriptionField) as string
+                        )
+                            .toLowerCase()
+                            .includes(searchText);
+                    }
+
+                    return includedInSearch;
+                })
+            );
+        }
+    }, [modelList, searchText]);
 
     return (
         <div>
-            {props.list &&
-                props.list.length > 0 &&
-                props.list.map(
-                    (model: TBaseModel, i: number): ReactElement => {
-                        const isSelected: boolean =
-                            props.selectedItems.filter(
-                                (selectedItem: TBaseModel) => {
-                                    return (
-                                        selectedItem._id?.toString() ===
-                                        model._id?.toString()
-                                    );
-                                }
-                            ).length > 0;
-
-                        return (
-                            <div
-                                key={i}
-                                onClick={() => {
-                                    props.onClick(model);
-                                }}
-                                className={`cursor-pointer mt-2 mb-2 relative flex items-center space-x-3 rounded-lg border border-gray-300 bg-white px-6 py-5 shadow-sm focus-within:ring-2 focus-within:ring-pink-500 focus-within:ring-offset-2 hover:border-gray-400 ${isSelected ? 'ring ring-indigo-500' : ''
-                                    }`}
-                            >
-                                <div className="min-w-0 flex-1">
-                                    <div className="focus:outline-none">
-                                        <span
-                                            className="absolute inset-0"
-                                            aria-hidden="true"
-                                        ></span>
-                                        {props.headerField &&
-                                            typeof props.headerField ===
-                                            Typeof.String && (
-                                                <p className="text-sm font-medium text-gray-300">
-                                                    {
-                                                        model.getValue(
-                                                            props.headerField as string
-                                                        ) as string
-                                                    }
-                                                </p>
-                                            )}
-
-                                        {props.headerField &&
-                                            typeof props.headerField ===
-                                            'function' &&
-                                            props.headerField(model)}
-                                        <p className="text-sm font-medium text-gray-900">
-                                            {
-                                                model.getValue(
-                                                    props.titleField
-                                                ) as string
-                                            }
-                                        </p>
-                                        {props.descriptionField && (
-                                            <p className="truncate text-sm text-gray-500">
-                                                {
-                                                    model.getValue(
-                                                        props.descriptionField
-                                                    ) as string
-                                                }
-                                            </p>
-                                        )}
-                                    </div>
-                                </div>
-                            </div>
-                        );
-                    }
+            <div>
+                {!isLoading && !error && props.isSearchEnabled && (
+                    <div className="p-2">
+                        <Input
+                            placeholder="Search..."
+                            onChange={(value: string) => {
+                                setSearchText(value);
+                            }}
+                        />
+                    </div>
                 )}
+            </div>
+            <div className="max-h-96 mb-5 overflow-y-auto p-2">
+                {error ? <ErrorMessage error={error} /> : <></>}
+                {isLoading ? <ComponentLoader /> : <></>}
+
+                {!isLoading && searchedList.length === 0 ? (
+                    <ErrorMessage
+                        error={
+                            searchText
+                                ? 'No items match your search'
+                                : props.noItemsMessage || 'No items found.'
+                        }
+                    />
+                ) : (
+                    <></>
+                )}
+
+                <StaicModelList<WorkflowVariable>
+                    list={searchedList}
+                    headerField={props.headerField}
+                    descriptionField={props.descriptionField}
+                    titleField={props.titleField}
+                    selectedItems={selectedList}
+                    onClick={(model: TBaseModel) => {
+                        if (props.selectMultiple) {
+                            // if added to the list, then remove or add to list
+                            const isSelected: boolean =
+                                selectedList.filter(
+                                    (selectedItem: TBaseModel) => {
+                                        return (
+                                            selectedItem._id?.toString() ===
+                                            model._id?.toString()
+                                        );
+                                    }
+                                ).length > 0;
+                            if (isSelected) {
+                                // remove the item.
+                                setSelectedList(
+                                    selectedList.filter(
+                                        (i: TBaseModel) => {
+                                            return (
+                                                i._id?.toString() !==
+                                                model._id?.toString()
+                                            );
+                                        }
+                                    )
+                                );
+                            } else {
+                                setSelectedList([
+                                    ...selectedList,
+                                    { ...model },
+                                ]);
+                            }
+                        } else {
+                            setSelectedList([{ ...model }]);
+                        }
+                    }}
+
+                />
+            </div>
         </div>
     );
 };
